requirements.txt
Flask==2.3.3
Markdown==3.10.2
python-frontmatter==1.1.0
PyGithub==2.8.1
requests==2.32.3
python-dotenv==1.2.1
supabase==2.28.0
argon2-cffi==25.1.0
bleach==6.1.0
google-genai>=1.0.0
Flask-Limiter==4.1.1
Flask-WTF==1.2.2
PyYAML==6.0.2
